requirements.txt
# Core LangGraph and LangChain dependencies
langgraph>=0.2.60
langchain>=0.1.0
langchain-core>=0.1.0
langchain-groq>=0.2.2

# LLM Provider clients
groq>=0.13.1
openai>=1.12.0
anthropic>=0.25.0

# Core dependencies
pydantic>=2.10.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0

# Additional utilities
asyncio>=3.4.3
json5>=0.9.0
typing-extensions>=4.8.0

# Development and testing (optional)
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0

# Performance optimization
uvloop>=0.19.0  # For async performance on Linux/macOS
aiohttp>=3.8.0  # For async HTTP requests

# Logging and monitoring
structlog>=23.0.0
python-json-logger>=2.0.0

# Configuration management
hydra-core>=1.3.0
omegaconf>=2.3.0

# Security
cryptography>=41.0.0
python-jose>=3.3.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Caching (optional for performance)
redis>=5.0.0
diskcache>=5.6.0

# WhatsApp integration (when ready)
flask>=2.3.0
gunicorn>=21.0.0
requests-toolbelt>=1.0.0

# Additional LangChain integrations
langchain-community>=0.0.25
langchain-experimental>=0.0.25

# Type checking
mypy>=1.5.0
types-requests>=2.31.0